% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox_area}
\alias{bbox_area}
\title{Functions for calculating with bbox
These functions can calculate various metrix of bbox}
\usage{
bbox_area(bbox)
}
\arguments{
\item{bbox}{character vector of bounding boxes to pad}
}
\value{
a vector of validated bboxes
}
\description{
Functions for calculating with bbox
These functions can calculate various metrix of bbox
}
\examples{
bbox_area("100 100 200 200")

}
