% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox_reset}
\alias{bbox_reset}
\title{Functions for updating certain coordinates of bbox
These functions can modify a vector of bbox}
\usage{
bbox_reset(bbox, x1 = NULL, y1 = NULL, x2 = NULL, y2 = NULL)
}
\arguments{
\item{bbox}{character vector of bounding boxes to update}

\item{x1}{scalar or numeric vector to update bbox with}

\item{y1}{scalar or numeric vector to update bbox with}

\item{x2}{scalar or numeric vector to update bbox with}

\item{y2}{scalar or numeric vector to update bbox with}
}
\value{
a vector of updated and validated bboxes
}
\description{
Functions for updating certain coordinates of bbox
These functions can modify a vector of bbox
}
\examples{
bbox_reset(bbox=c("100 100 200 200", "300 400 500 600"), x2=800)

}
