% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox_pad_width}
\alias{bbox_pad_width}
\alias{bbox_pad_height}
\title{Functions for padding bbox
These functions can "pad" (increase size of) bbox}
\usage{
bbox_pad_width(bbox, word, n = 1, side = "both")

bbox_pad_height(bbox, word, n = 0.5, side = "both")
}
\arguments{
\item{bbox}{character vector of bounding boxes to pad}

\item{word}{character vector of words contained in bboxes}

\item{n}{integer number of symbols to add}

\item{side}{"left", "right" or "both" which side to pad}

\item{side}{"up", "down" or "both" which side to pad}
}
\value{
a vector of validated bboxes
}
\description{
Functions for padding bbox
These functions can "pad" (increase size of) bbox
}
\examples{
bbox_pad_width("0 0 10 20", "There")
bbox_pad_height("0 0 10 20", "There")
}
