% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox_slice_x}
\alias{bbox_slice_x}
\alias{bbox_slice_y}
\alias{bbox_slice_xy}
\title{Functions for horizontal and vertical slicing of bbox columns
These functions are not vectorized and should be used to compute}
\usage{
bbox_slice_x(bbox, x)

bbox_slice_y(bbox, y)

bbox_slice_xy(bbox, x, y)
}
\arguments{
\item{bbox}{string or character vector of bbox'es}

\item{x, y}{coordinates to sclice the bouding box at}
}
\value{
list of bbox'es appliccable for particular split
}
\description{
Functions for horizontal and vertical slicing of bbox columns
These functions are not vectorized and should be used to compute
}
\examples{
bbox_slice_x("0 0 100 200", 80)
bbox_slice_x(c("0 0 100 200", "100 100 200 200"), c(80, 150))

bbox_slice_y("0 0 100 200", 120)
bbox_slice_y(c("0 0 100 200", "100 100 200 200"), c(120,150))
bbox_slice_xy("0 0 100 200", 50, 100)
bbox_slice_xy(c("0 0 100 200", "100,100, 200, 200"), c(50, 150), c(100, 150))
}
