% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox_to_geometry}
\alias{bbox_to_geometry}
\title{Convert bbox to magick-geometry}
\usage{
bbox_to_geometry(x)
}
\arguments{
\item{x}{character vector with bounding boxes (top-left coordinates). Expects four integers separated by comma or space}
}
\value{
character vector of magick-compliant geometry string representing bounding box area in the format  "width x height +x_off +y_off" (https://www.imagemagick.org/Magick++/Geometry.html)
}
\description{
Convert bbox to magick-geometry
}
\examples{
# area from c(0,0) to c(100,200) counting from top left
bbox_to_geometry("0 0 100 200")

}
\seealso{
\link[magick:geometry]{magick::geometry} for converting integers (four individual columns) to geometry
}
