% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox_is_valid}
\alias{bbox_is_valid}
\alias{bbox_validate}
\title{Functions for validating bbox
These functions can check whether specified bbox is valid, i.e. x1 <= x2 and y1 <= y2}
\usage{
bbox_is_valid(bbox)

bbox_validate(bbox)
}
\arguments{
\item{bbox}{character vector bounding boxes to validate}
}
\value{
a vector of logical values

original vector with NA for invalid bboxes
}
\description{
Functions for validating bbox
These functions can check whether specified bbox is valid, i.e. x1 <= x2 and y1 <= y2
}
\examples{
bbox_is_valid("0 0 100 200")
bbox_validate(c("5,4,6,3", "1,1,5,6"))
}
