% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{separate_bbox_cols}
\alias{separate_bbox_cols}
\title{Create four individual columns out of each bbox column and add to tidy tibble.}
\usage{
separate_bbox_cols(df, bbox_cols = NULL, drop = FALSE)
}
\arguments{
\item{df}{tidy hocr tibble, as produced by tidy_tesseract}

\item{bbox_cols}{character vector listing all the bbox columns you want to transform, if ommitted till transform all bbox columns}

\item{drop}{logical, indicating whether original bbox column(s) shoulc be dropped}
}
\value{
tidy tibble with four column added per every transformed bbox column
}
\description{
Create four individual columns out of each bbox column and add to tidy tibble.
}
\examples{
\dontrun{
library(tesseract)

ocr("text.png", HOCR=TRUE) \%>\%
  hocr_parse() \%>\%
  tidy_tesseract() \%>\%
  separate_bbox_cols(drop=TRUE)
}
}
